[chore] Removed Integrity Checking
This was deemed to be not really that useful. Instead I implemented to check if the file
is 0 bytes or does not exist to download the page. This has drawbacks but it is up
for the user to perform manual checks to the download to determine whether they got a
correct file or not.